requirements.txt
librosa
soundfile
scikit-learn>=1.3.2
moviepy
numpy
gradio
faster-whisper==0.10.0
ffmpeg-python==0.2.0
transformers==4.36.2
